currency: Migrate LongCurrencyFormatter and add try_new_long
