+++ components/experimental/src/dimension/currency/mod.rs
@@ -10,11 +10,9 @@ pub mod format;
 pub mod formatter;
 pub mod long_compact_format;
 pub mod long_compact_formatter;
-pub mod long_format;
-pub mod long_formatter;
 pub mod options;
 
 /// A currency code, such as "USD" or "EUR".
-#[derive(Clone, Copy, Debug)]
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
 #[allow(clippy::exhaustive_structs)] // newtype
 pub struct CurrencyCode(pub TinyAsciiStr<3>);
